workflows/requirements.txt
-i https://pypi.org/simple
adal==1.2.7
aenum==3.1.11
aiohttp==3.8.1
aiosignal==1.3.1
alembic==1.9.1
amqp==5.1.1
ansiwrap==0.8.4
anyio==3.6.2
apache-airflow==2.5.0
apache-airflow-providers-amazon==6.2.0
apache-airflow-providers-celery==3.1.0
apache-airflow-providers-cncf-kubernetes==5.0.0
apache-airflow-providers-common-sql==1.3.1
apache-airflow-providers-discord==3.1.0
apache-airflow-providers-ftp==3.2.0
apache-airflow-providers-http==4.1.0
apache-airflow-providers-imap==3.1.0
apache-airflow-providers-papermill==3.1.0
apache-airflow-providers-postgres==5.3.1
apache-airflow-providers-redis==3.1.0
apache-airflow-providers-slack==7.1.0
apache-airflow-providers-sqlite==3.3.1
apache-airflow-providers-telegram==3.1.0
apispec[yaml]==3.3.2
appnope==0.1.3
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
astro-sdk-python==1.3.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
awswrangler==2.18.0
azure-core==1.26.1
azure-datalake-store==0.0.52
azure-storage-blob==12.14.1
babel==2.11.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==23.1a1
bleach==5.0.1
blinker==1.5
boto3==1.26.41
botocore==1.29.41
cachecontrol==0.12.11
cachelib==0.9.0
cachetools==5.2.0
cattrs==22.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.1.2
commonmark==0.9.1
configupdater==3.1.1
connexion[flask,swagger-ui]==2.14.1
contourpy==1.0.6
cron-descriptor==1.2.32
croniter==1.3.8
cryptography==39.0.0
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
deprecated==1.2.13
dill==0.3.6
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
flask==2.2.2
flask-appbuilder==4.1.4
flask-babel==2.0.0
flask-caching==2.0.1
flask-jwt-extended==4.4.4
flask-login==0.6.2
flask-session==0.4.0
flask-sqlalchemy==2.5.1
flask-wtf==1.0.1
flower==1.2.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
gcsfs==2022.11.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
graphviz==0.20.1
gremlinpython==3.6.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.2
humanize==4.4.0
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jinja2==3.1.2
jmespath==1.0.1
jsonpath-ng==1.5.3
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.4.8
jupyter-console==6.4.4
jupyter-core==5.1.2
jupyter-events==0.5.0
jupyter-server==2.0.6
jupyter-server-terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kombu==5.2.4
kubernetes==23.6.0
lazy-object-proxy==1.8.0
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==4.9.2
mako==1.2.4
markdown==3.4.1
markdown-it-py==2.1.0
markupsafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.4
mpmath==1.2.1
msgpack==1.0.4
msrest==0.7.1
multidict==6.0.4
mypy-boto3-appflow==1.26.32
mypy-boto3-rds==1.26.39
mypy-boto3-redshift-data==1.26.30
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.23.3
oauthlib==3.2.2
openpyxl==3.0.10
opensearch-py==2.0.1
packaging==22.0
pandas==1.5.1
pandocfilters==1.5.0
papermill[all]==2.4.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.3
pendulum==2.1.2
pexpect==4.8.0
pg8000==1.29.4
pickleshare==0.7.5
pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
ply==3.11
prison==0.2.1
progressbar2==4.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pygithub==1.57
pygments==2.14.0
pygraphviz==1.10
pyjwt==2.6.0
pymysql==1.0.2
pynacl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-daemon==2.3.2
python-dateutil==2.8.2
python-frontmatter==1.0.0
python-json-logger==2.0.4
python-nvd3==0.15.0
python-slugify==7.0.0
python-telegram-bot==20.0
python-utils==3.4.5
pytz==2022.7
pytzdata==2020.1
pyyaml==6.0
pyzmq==24.0.1
qtconsole==5.4.0
qtpy==2.3.0
redis==3.5.3
redshift-connector==2.0.909
requests==2.28.1
requests-aws4auth==1.1.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986[idna2008]==1.5.0
rfc3986-validator==0.1.1
rich==13.0.0
rsa==4.9
s3transfer==0.6.0
scipy==1.9.3
scramp==1.4.4
scrapbook[all]==0.5.0
send2trash==1.8.0
setproctitle==1.3.2
setuptools==65.6.3
six==1.16.0
slack-sdk==3.19.5
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlalchemy==1.4.45
sqlalchemy-jsonfield==1.0.1.post0
sqlalchemy-redshift==0.8.12
sqlalchemy-utils==0.39.0
sqlparse==0.4.3
stack-data==0.6.2
statsd==4.0.1
statsmodels==0.13.5
swagger-ui-bundle==0.0.9
sympy==1.11.1
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.1.1
terminado==0.17.1
text-unidecode==1.3
textwrap3==0.9.2
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typing-extensions==4.4.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
watchtower==2.0.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
wtforms==3.0.1
yarl==1.8.2
zipp==3.11.0
